backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
llama-index
llama-index-vector-stores-chroma
chromadb
python-dotenv
